# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: WECTI API
  description: >
    API central do sistema de controle de acesso a palestras e eventos.
    O site (aluno e admin/professor) consome exclusivamente esta API -
    nenhum cliente acessa o banco direto. Nao ha app Android dedicado: o
    check-in/check-out é feito via QR code lido pela câmera nativa do
    celular do aluno. Este e um RASCUNHO inicial para alinhar o time
    antes da Fase 1; campos e regras podem mudar conforme o
    desenvolvimento avança.
  version: 0.1.0-draft
servers:
  - url: http://localhost:8080/v1
    description: Ambiente local de desenvolvimento
  - url: https://api.wecti.com.br/v1
    description: Produção (placeholder, ajustar quando definirem o domínio)
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Autentica um usuário (aluno, professor ou admin)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha]
              properties:
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  format: password
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '401':
          description: Credenciais inválidas
  /auth/registrar:
    post:
      tags: [Auth]
      summary: Cadastro público de aluno ("Primeiro acesso? Crie sua conta")
      description: >
        Sempre cria um usuário com perfil ALUNO - não existe campo
        "perfil" no corpo da requisição, então não tem como quem se
        autocadastra virar Professor ou Admin por essa via. Cadastro de
        Professor continua sendo feito só pelo Admin, em POST /usuarios.
        Já devolve token (igual ao login), pra entrar direto no sistema.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, email, senha, rgm]
              properties:
                nome:
                  type: string
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  format: password
                  minLength: 6
                rgm:
                  type: string
                  pattern: '^\d{8}$'
                curso:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos (email malformado, RGM fora do formato, senha curta)
        '409':
          description: Já existe um usuário com este email ou RGM
  /auth/redefinir-senha:
    post:
      tags: [Auth]
      summary: "\"Esqueci minha senha\" - redefine sem link por email"
      description: >
        Sem token/link por email de propósito - não há servidor de email
        configurado no projeto, e o cliente confirmou que não precisa de
        autenticação extra. A identidade é confirmada com o RGM (se
        perfil = aluno) ou CPF (se perfil = professor) já cadastrados -
        batendo, a pessoa já define a senha nova na mesma chamada. Serve
        também pro Professor (sempre cadastrado pelo Admin com uma senha
        provisória que ninguém sabe) definir a primeira senha de verdade.
        Não cobre reset de conta Admin (sem RGM/CPF) - por design.
        Já devolve token (igual ao login), pra entrar direto no sistema.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, identificador, nova_senha]
              properties:
                email:
                  type: string
                  format: email
                identificador:
                  type: string
                  description: RGM (8 dígitos, aluno) ou CPF (11 dígitos, professor)
                nova_senha:
                  type: string
                  format: password
                  minLength: 6
      responses:
        '200':
          description: Senha redefinida
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos (email malformado, senha curta)
        '401':
          description: Email + RGM/CPF não conferem com nenhum cadastro

  # ---------------------------------------------------------------------
  # Usuarios
  # ---------------------------------------------------------------------
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Lista usuários paginados (somente admin)
      description: >
        Paginado de propósito - a lista de usuários cresce muito quando o
        sistema é liberado pros alunos, devolver tudo de uma vez não
        escala. "busca" filtra por nome, email, RGM ou CPF (contém, sem
        diferenciar maiúsculas/minúsculas no nome/email).
      parameters:
        - name: perfil
          in: query
          schema:
            $ref: '#/components/schemas/Perfil'
        - name: busca
          in: query
          schema:
            type: string
        - name: page
          in: query
          description: Página (começando em 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Página de usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  conteudo:
                    type: array
                    items:
                      $ref: '#/components/schemas/Usuario'
                  pagina:
                    type: integer
                  tamanho:
                    type: integer
                  total_elementos:
                    type: integer
                  total_paginas:
                    type: integer
    post:
      tags: [Usuarios]
      summary: Cadastra aluno ou convidado avulso
      description: >
        Usado pelo admin/professor direto pelo site quando precisa
        cadastrar alguém na hora, na porta do evento.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovoUsuario'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
  /usuarios/{usuarioId}:
    parameters:
      - name: usuarioId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Usuarios]
      summary: Atualiza um usuário (admin ou professor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovoUsuario'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '409':
          description: Email ou RGM já usado por outro usuário
    delete:
      tags: [Usuarios]
      summary: Exclui um usuário (somente admin)
      responses:
        '204':
          description: Usuário excluído
        '422':
          description: Usuário possui inscrições e não pode ser excluído
...
  /eventos/{eventoId}/checkin-sessoes:
    parameters:
      - name: eventoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Checkins]
      summary: Gera uma sessão de check-in/check-out pro evento (admin ou professor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaSessaoCheckin'
      responses:
        '201':
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessaoCheckin'
  /checkin-sessoes/{sessaoId}/qrcode:
    parameters:
      - name: sessaoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Checkins]
      summary: Imagem do QR code da sessão, pra projetar na tela (admin ou professor)
      description: >
        O QR codifica a URL de confirmação
        (`<frontend>/checkin/confirmar/{sessaoId}`) - a camera do celular
        do aluno abre direto, sem precisar de app.
      responses:
        '200':
          description: Imagem do QR code
          content:
            image/png:
              schema:
                type: string
                format: binary
        '422':
          description: Sessão já expirada
  /checkin-sessoes/{sessaoId}/confirmar:
    parameters:
      - name: sessaoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Checkins]
      summary: Aluno confirma a própria entrada/saída ao escanear o QR
      responses:
        '200':
          description: Check-in ou check-out confirmado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checkin'
        '404':
          description: QR code inválido, expirado, ou aluno não inscrito no evento
        '409':
          description: Check-in já realizado
        '422':
          description: >
            QR expirado, inscrição cancelada, ou tentativa de check-out
            sem ter feito check-in antes

  # ---------------------------------------------------------------------
  # Certificados
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Perfil:
      type: string
      # Documentado em maiusculas porque e assim que o Jackson serializa o
      # enum Java por padrao (Usuario.perfil, NovoUsuario.perfil e a claim
      # "perfil" do JWT) - o rascunho original documentava em minusculas,
      # mas nao era isso que a API implementada realmente devolvia/esperava.
      enum: [ADMIN, PROFESSOR, ALUNO]
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          $ref: '#/components/schemas/Perfil'
        rgm:
          type: string
          pattern: '^\d{8}$'
          nullable: true
          description: >
            Registro Geral do Aluno (RGM) - identificador acadêmico da
            faculdade, com 8 dígitos. Obrigatório e único quando
            perfil = aluno; não se aplica a admin/professor.
        cpf:
          type: string
          pattern: '^\d{11}$'
          nullable: true
          description: >
            CPF do professor, 11 dígitos. Obrigatório e único quando
            perfil = professor; não se aplica a admin/aluno.
        curso:
          type: string
          nullable: true
          description: >
            Curso do aluno (ex.: "Ciência da Computação") - confirmado
            com o professor que é só informativo, pra saber quais cursos
            estão participando das palestras. Não afeta pontuação,
            elegibilidade pra inscrição nem nenhuma outra regra de
            negócio - por isso texto livre, opcional, sem unicidade.
            Não se aplica a admin/professor.
    NovoUsuario:
      type: object
      required: [nome, email, perfil]
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          $ref: '#/components/schemas/Perfil'
        rgm:
          type: string
          pattern: '^\d{8}$'
          nullable: true
          description: >
            Obrigatório quando perfil = aluno. O OpenAPI puro não expressa
            "obrigatório somente se outro campo tiver determinado valor",
            então essa regra precisa ser validada no backend.
        cpf:
          type: string
          pattern: '^\d{11}$'
          nullable: true
          description: >
            Obrigatório quando perfil = professor, pela mesma razão do rgm
            acima - validado no backend.
        curso:
          type: string
          nullable: true
          description: Opcional, só informativo (ver descrição em Usuario.curso).
...
    Checkin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inscricao_id:
          type: string
          format: uuid
        entrada:
          type: string
          format: date-time
        saida:
          type: string
          format: date-time
          nullable: true
        percentual_presenca:
          type: number
          format: float
          description: Calculado quando há saída registrada
    NovaSessaoCheckin:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
          enum: [ENTRADA, SAIDA]
    SessaoCheckin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        evento_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [ENTRADA, SAIDA]
        criada_em:
          type: string
          format: date-time
        expira_em:
          type: string
          format: date-time
          description: >
            Validade do QR code (6 horas após a criação) - depois disso
            o endpoint de confirmação passa a responder 422.
...